+++ libscidavis/src/Integration.cpp
@@ -157,8 +157,8 @@ QString Integration::logInfo()
 	// using GSL to find maximum value of data set
 	gsl_vector *aux = gsl_vector_alloc(d_n);
 	for(unsigned i=0; i < d_n; i++)
-		gsl_vector_set (aux, i, d_y[i]);
-	int maxID=gsl_vector_max_index (aux);
+        gsl_vector_set (aux, i, d_y[i]);
+    int maxID = static_cast<int>(gsl_vector_max_index (aux));
 	gsl_vector_free (aux);
 
 	// calculate result
